[FIX-IMP] product_version modules: posibility to add product version in new picking + move menu of purchase version + product version onchange fix + AttributeError fix

+++ product_variant_custom_purchase/models/purchase.py
@@ -78,6 +78,10 @@ class PurchaseOrderLine(models.Model):
     @api.multi
     @api.onchange('product_tmpl_id')
     def onchange_product_template(self):
+        if self.product_tmpl_id and self.product_id.product_tmpl_id == self.product_tmpl_id:
+            return {'domain': {'product_id':
+                               [('product_tmpl_id', '=',
+                                 self.product_tmpl_id.id)]}}
         self.ensure_one()
         self.product_attribute_ids = \
             self._delete_product_attribute_ids()
@@ -100,11 +104,16 @@ class PurchaseOrderLine(models.Model):
 
     @api.onchange('product_id')
     def onchange_product_id(self):
+        p1 = self.product_id
         result = super().onchange_product_id()
-        self.custom_value_ids = self._delete_custom_lines()
-        self.product_attribute_ids = self._delete_product_attribute_ids()
+        if not self.product_tmpl_id and self.product_id:
+            self.product_tmpl_id = self.product_id.product_tmpl_id
+            self.product_id = p1
         if self.product_id:
-            product = self.product_id
+            self.custom_value_ids = self._delete_custom_lines()
+            self.product_attribute_ids = self._delete_product_attribute_ids()
+            product = p1
+
             self.product_attribute_ids = \
                 product._get_product_attributes_values_dict()
